begin refactoring and cleaning up redux code
