--- src/common/Hooks.ts
@@ -1,7 +1,0 @@
-import { TypedUseSelectorHook, useDispatch, useSelector } from 'react-redux'
-import { AppDispatch, RootState } from '../redux/ReduxStore'
-
-// Use instead of plain `useDispatch` and `useSelector` 
-// - allows typing and was suggested by react-redux
-export const useAppDispatch: () => AppDispatch = useDispatch
-export const useAppSelector: TypedUseSelectorHook<RootState> = useSelector
